Instantiating Service
This code instantiate a service to a component to share the logic. There is a problem with that approach. If you wish to call the service on another component you need to instantiate it again and your not working on the same instance so maybe the data your working on is not the same or updated.